refactor: rename remote_package module
