+++ rocks-bin/src/unpack.rs
@@ -3,7 +3,7 @@ use std::{fs::File, io::Cursor, path::PathBuf};
 use clap::Args;
 use eyre::Result;
 use indicatif::MultiProgress;
-use rocks_lib::{config::Config, remote_package::PackageReq};
+use rocks_lib::{config::Config, package::PackageReq};
 
 #[derive(Args)]
 pub struct Unpack {
