+++ rocks-lib/resources/test/sample-tree/5.1/lock.json
@@ -1,1 +1,1 @@
-{"version":"1.0.0","rocks":{"d7164c4921869877c7f29fce3f14b8ea78e0aec0d0c36123a3a920a894785ea5":{"name":"neorg","version":"8.0.0-1","pinned":false,"dependencies":["48ec344951668eca0e0a4ff284d804a11e4e709194df3191a72ed8fac89cf2e0"],"constraint":null,"hashes":{"rockspec":"sha256-uU0nuZNNPgilLlLX2n2r+sSE7+N6U4DukIj3rOLvzek=","source":"sha256-uU0nuZNNPgilLlLX2n2r+sSE7+N6U4DukIj3rOLvzek="}},"aa0a5bcd396f3b8e59fa9dd8059a06c3bf4952f48473214d96fc46895edb59f0":{"name":"neorg","version":"8.8.1-1","pinned":false,"dependencies":["48ec344951668eca0e0a4ff284d804a11e4e709194df3191a72ed8fac89cf2e0"],"constraint":null,"hashes":{"rockspec":"sha256-uU0nuZNNPgilLlLX2n2r+sSE7+N6U4DukIj3rOLvzek=","source":"sha256-uU0nuZNNPgilLlLX2n2r+sSE7+N6U4DukIj3rOLvzek="}},"48ec344951668eca0e0a4ff284d804a11e4e709194df3191a72ed8fac89cf2e0":{"name":"lua-cjson","version":"2.1.0-1","pinned":false,"dependencies":[],"constraint":null,"hashes":{"rockspec":"sha256-uU0nuZNNPgilLlLX2n2r+sSE7+N6U4DukIj3rOLvzek=","source":"sha256-uU0nuZNNPgilLlLX2n2r+sSE7+N6U4DukIj3rOLvzek="}}},"entrypoints":["d7164c4921869877c7f29fce3f14b8ea78e0aec0d0c36123a3a920a894785ea5","aa0a5bcd396f3b8e59fa9dd8059a06c3bf4952f48473214d96fc46895edb59f0"]}
+{"version":"1.0.0","rocks":{"aa0a5bcd396f3b8e59fa9dd8059a06c3bf4952f48473214d96fc46895edb59f0":{"name":"neorg","version":"8.8.1-1","pinned":false,"dependencies":["48ec344951668eca0e0a4ff284d804a11e4e709194df3191a72ed8fac89cf2e0"],"constraint":null,"hashes":{"rockspec":"sha256-uU0nuZNNPgilLlLX2n2r+sSE7+N6U4DukIj3rOLvzek=","source":"sha256-uU0nuZNNPgilLlLX2n2r+sSE7+N6U4DukIj3rOLvzek="}},"d7164c4921869877c7f29fce3f14b8ea78e0aec0d0c36123a3a920a894785ea5":{"name":"neorg","version":"8.0.0-1","pinned":false,"dependencies":["48ec344951668eca0e0a4ff284d804a11e4e709194df3191a72ed8fac89cf2e0"],"constraint":null,"hashes":{"rockspec":"sha256-uU0nuZNNPgilLlLX2n2r+sSE7+N6U4DukIj3rOLvzek=","source":"sha256-uU0nuZNNPgilLlLX2n2r+sSE7+N6U4DukIj3rOLvzek="}},"48ec344951668eca0e0a4ff284d804a11e4e709194df3191a72ed8fac89cf2e0":{"name":"lua-cjson","version":"2.1.0-1","pinned":false,"dependencies":[],"constraint":null,"hashes":{"rockspec":"sha256-uU0nuZNNPgilLlLX2n2r+sSE7+N6U4DukIj3rOLvzek=","source":"sha256-uU0nuZNNPgilLlLX2n2r+sSE7+N6U4DukIj3rOLvzek="}}},"entrypoints":["aa0a5bcd396f3b8e59fa9dd8059a06c3bf4952f48473214d96fc46895edb59f0","d7164c4921869877c7f29fce3f14b8ea78e0aec0d0c36123a3a920a894785ea5"]}
