+++ rocks-lib/src/lib.rs
@@ -5,8 +5,8 @@ pub mod lockfile;
 pub mod lua_installation;
 pub mod manifest;
 pub mod operations;
+pub mod package;
 pub mod progress;
 pub mod project;
-pub mod remote_package;
 pub mod rockspec;
 pub mod tree;
